requirements.txt
arxiv==2.2.0
openreview==0.0.2
openreview_py==1.51.2
pandas==2.3.3
pdfminer==20191125
pdfminer_six==20250506
psycopg2==2.9.10
Requests==2.32.5
tqdm==4.67.1
APScheduler>=3.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0  # if using PostgreSQL
requests>=2.31.0
